Exclusive thread possible fix for Linux
The deadlock is possibly caused by the thread not being joined, or because of an implementation difference with phtreads.

So this fix adds a call to release the background thread when terminating it and then also joins the thread to await for the completion of the background thread.

As before this works fine under Win32 (x64)