Compositor: Allow final render to texture
Now it is possible to render the result of an entire Compositor pipeline to a
texture, instead of the screen. However, the FBOs aren't shared between passes
rendering to this target texture. In passes rendering directly to the screen,
the screen framebuffer is shared. OSG doesn't provide an easy way of sharing FBOs.