Fix the Bermuda Triangle -ve scenery tex coords
In some cases TerraGear generates scenery triangles with
negative texture coordinates.

With negative tex coords, FG generated out of bounds texture
coordinates for the object and vegetation masks and crashed.

This fix uses some explicit casts to ensure that the texture
coordinates are always positive.  This makes FG more robust
but doesn't fix the underlying problem which is that the texture
coorinates are wrong.