SGProgram: Validate per-context shader state
SGProgram::apply() assumes that every shader in a program which failed
to apply has a valid PerContextShader returned by osg::Shader::getPCS().
However this may not be the case if the program contains conditional
shaders, which depend on some define via a "#pragma requires". A shader
compilation or link error in such a program triggers a segfault.

Add validation to skip the shader if no PerContextShader object exists.

An example of this is conditional geometry shaders for geometry shader
based multiview rendering. Even without that yet, lets make sure this
doesn't bite again.