Convert files in simgear to use SPDX.
As per the discussion on SourceForge ticket #1940, as well as the
long-standing and ongoing effort to codify SimGear as LGPL-licensed,
this PR changes GPL-licensed files which were supposed to have been
changed over to LGPL.

There is some code that is BSD 3-clause licensed. The license on those
files has been respected and remains unchanged, its just switched to using
an SPDX identifier instead of the traditional preamble.

Additionally where convenient some adjustments were made to Doxygen tags.
I have attempted not to change the content or meaning of the documentation
where possible, just simply to 'doxygenify' it.