GLcore: Compositor: Use another technique to enable/disable rendering of passes
Previously we used cull masks to choose whether to render a pass or not.
This approach leaks some draw calls, so we will attempt to use graphics
contexts to toggle rendering. Setting it to NULL on an osg::Camera
effectively disables rendering.

The new approach might have unintended consequences (specially when
interacting with threading), but it is worth a shot.