GLcore: Fix instancing
OSG's VertexAttribDivisor calls glVertexAttribDivisor() before the
VAO containing the instanced geometry is bound. This ordering causes
the divisor to affect the VAO that was bound before, which is incorrect.

Our OSG fork now contains a fix that treats the vertex divisor as a
property of osg::Array, and correctly calls glVertexAttribDivisor() after
the VAO containing the instanced geometry is bound.

See OSG commit 7531c2b8.