WS30: Use tessellation shaders
Geometry is no longer explicitly defined on the CPU. A list of patches
is sent to the GPU and tessellated there. Normals are calculated on the
fly by the shaders and don't need to be sent as a vertex attribute.

The tessellation algorithm requires tiles to be continuous, i.e. we need
to obtain elevation info from adjacent tiles for edge vertices.

The sea plane has been commented out for now.

See related fgdata commit bb949d352.